@@ -39,6 +39,9 @@
     "@radix-ui/react-toggle-group": "1.1.1",
     "@radix-ui/react-tooltip": "1.1.6",
     "axios": "latest",
+    "badge": "link:@/components/ui/badge",
+    "button": "link:@/components/ui/button",
+    "card": "link:@/components/ui/card",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "1.0.4",
@@ -49,20 +52,23 @@
     "geist": "^1.3.1",
     "input-otp": "1.4.1",
     "lucide-react": "^0.454.0",
-    "next": "15.2.4",
+    "next": "15.4.4",
     "next-themes": "latest",
-    "react": "^19",
+    "progress": "link:@/components/ui/progress",
+    "react": "^19.1.0",
     "react-day-picker": "9.8.0",
-    "react-dom": "^19",
+    "react-dom": "^19.1.0",
     "react-hook-form": "^7.54.1",
     "react-hot-toast": "latest",
     "react-markdown": "latest",
     "react-resizable-panels": "^2.1.7",
     "recharts": "2.15.0",
+    "separator": "link:@/components/ui/separator",
     "sonner": "^1.7.1",
     "tailwind-merge": "^2.5.5",
     "tailwindcss-animate": "^1.0.7",
     "vaul": "^0.9.6",
+    "web3": "^4.16.0",
     "zod": "^3.24.1"
   },
   "devDependencies": {
