@@ -10,9 +10,10 @@ import { ThemeProvider } from "@/components/theme-provider"
 const inter = Inter({ subsets: ["latin"] })
 
 export const metadata: Metadata = {
-  title: "OpenLearnX - Decentralized Adaptive Learning",
-  description: "AI-powered adaptive testing with blockchain-secured credentials.",
-    generator: 'v0.dev'
+  title: "OpenLearnX - Comprehensive Learning Dashboard",
+  description: "AI-powered adaptive learning with blockchain integration, real-time analytics, and professional progress tracking.",
+  keywords: "learning, coding, blockchain, AI, analytics, professional development",
+  generator: 'OpenLearnX v2.0'
 }
 
 export default function RootLayout({
@@ -25,9 +26,22 @@ export default function RootLayout({
       <body className={inter.className}>
         <ThemeProvider attribute="class" defaultTheme="system" enableSystem disableTransitionOnChange>
           <AuthProvider>
-            <Navbar />
-            <main>{children}</main>
-            <Toaster position="top-right" />
+            <div className="min-h-screen bg-gradient-to-br from-slate-50 via-blue-50 to-indigo-100 dark:from-gray-900 dark:via-blue-900 dark:to-purple-900">
+              <Navbar />
+              <main className="transition-all duration-300">{children}</main>
+              <Toaster 
+                position="top-right" 
+                toastOptions={{
+                  duration: 4000,
+                  style: {
+                    background: 'rgba(17, 24, 39, 0.95)',
+                    color: '#fff',
+                    backdropFilter: 'blur(16px)',
+                    border: '1px solid rgba(75, 85, 99, 0.3)',
+                  },
+                }}
+              />
+            </div>
           </AuthProvider>
         </ThemeProvider>
       </body>
